% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grm_cal.R
\name{grm_cal}
\alias{grm_cal}
\title{Calculate GRM (Genetic Relationship Matrix)}
\usage{
grm_cal(freq_table, geno_table, weight_vec = NULL)
}
\arguments{
\item{freq_table}{A data frame containing allele frequencies.}

\item{geno_table}{A data frame containing genotype data.}

\item{weight_vec}{A numeric vector of weights (default: NULL).}
}
\value{
A computed GRM matrix.
}
\description{
This function computes a genetic relationship matrix based on genotype and frequency data.
}
\examples{
data(test_geno)
data(test_freq)
grm_cal(test_freq, test_geno)
}
